requirements.txt
streamlit>=1.28.0
pillow>=9.0.0
numpy==1.26.4
torch==2.2.2
torchvision==0.17.2
transformers==4.45.0
accelerate>=0.20.0
scipy>=1.10.0
psutil>=5.8.0
--index-url https://download.pytorch.org/whl/cpu